Add files via upload
Python Financial Analysis Capstone project in Jupyter Notebook - Python, Pandas, NumPy, Matplotlib, cvxopt, seaborn.

The purpose of this capstone is to practice accessing and downloading financial data, calculating financial statistics to understand the individual performance of an asset and how groups of assets are related, and how to allocate investments across assets for optimal performance.

Investment Criteria
The portfolio below is a combination of my own research and suggestions by professional asset managers who manage large portfolios. I have picked seven stocks based on the following criteria:

Technologically enabled
Going to transform the the world as we know it
Industries that are going to see incredible growth in next 5-10yrs:
Robotics
Artificial intelligence
Energy storage
Blockchain technology
Multiomic sequencing (lifescience) - is a research approach that integrates data from multiple "omics" levels (like genomics, transcriptomics, proteomics, etc.) to provide a more comprehensive understanding of biological systems and disease mechanisms. It reveals complex interactions and relationships within a cell or organism.
It is going to be a 5-10 year strategy that is for more risk-tolerant investor. I am expecting to see large standard deviations on some of them. We are at the gusp of the world changing drastically as a result of Gen AI becoming mainstream. The new technologies are developing within matter of weeks, not matter of years, therefore I felt it would make sense to consder the stocks listed below. Further reasoning and descriptions are in the attached PowerPoint document.

Stock analysis, risk, return, correlation heat map, efficient frontier, simulation of 5000 scenarios to find the optimal portfolio.